+++ packages/expo-screen-capture/src/ExpoScreenCapture.web.ts
@@ -1,5 +1,1 @@
-export default {
-  get name(): string {
-    return 'ExpoScreenCapture';
-  },
-};
+export default {};