+++ packages/expo-sensors/src/ExponentMagnetometerUncalibrated.web.ts
@@ -1,7 +1,4 @@
 export default {
-  get name(): string {
-    return 'ExponentMagnetometerUncalibrated';
-  },
   async isAvailableAsync(): Promise<boolean> {
     return false;
   },