+++ packages/expo-sms/src/ExpoSMS.ts
@@ -1,2 +1,1 @@
-import { requireNativeModule } from 'expo-modules-core';
-export default requireNativeModule('ExpoSMS');
+export default {} as any;