+++ packages/expo-updates/src/ExpoUpdates.web.ts
@@ -1,11 +1,5 @@
-import { Platform } from 'expo-modules-core';
-
 export default {
-  get name(): string {
-    return 'ExpoUpdates';
-  },
   async reload(): Promise<void> {
-    if (!Platform.isDOMAvailable) return;
-    window.location.reload(true);
+    if (typeof window !== 'undefined') window.location.reload(true);
   },
 };