+++ packages/expo-av/plugin/src/withAV.ts
@@ -1,9 +1,4 @@
-import {
-  AndroidConfig,
-  ConfigPlugin,
-  createRunOncePlugin,
-  withInfoPlist,
-} from 'expo/config-plugins';
+import { AndroidConfig, ConfigPlugin, IOSConfig, createRunOncePlugin } from 'expo/config-plugins';
 
 const pkg = require('expo-av/package.json');
 
@@ -13,13 +8,11 @@ const withAV: ConfigPlugin<{ microphonePermission?: string | false } | void> = (
   config,
   { microphonePermission } = {}
 ) => {
-  if (microphonePermission !== false) {
-    config = withInfoPlist(config, (config) => {
-      config.modResults.NSMicrophoneUsageDescription =
-        microphonePermission || config.modResults.NSMicrophoneUsageDescription || MICROPHONE_USAGE;
-      return config;
-    });
-  }
+  IOSConfig.Permissions.createPermissionsPlugin({
+    NSMicrophoneUsageDescription: MICROPHONE_USAGE,
+  })(config, {
+    NSMicrophoneUsageDescription: microphonePermission,
+  });
 
   return AndroidConfig.Permissions.withPermissions(
     config,