+++ packages/expo-barcode-scanner/plugin/src/withBarcodeScanner.ts
@@ -1,22 +1,20 @@
-import { AndroidConfig, ConfigPlugin, createRunOncePlugin } from 'expo/config-plugins';
+import { AndroidConfig, ConfigPlugin, IOSConfig, createRunOncePlugin } from 'expo/config-plugins';
 
 const pkg = require('expo-barcode-scanner/package.json');
 
-const CAMERA_USAGE = 'Allow $(PRODUCT_NAME) to access your camera';
-const MICROPHONE_USAGE = 'Allow $(PRODUCT_NAME) to access your microphone';
-
 const withBarcodeScanner: ConfigPlugin<
   {
-    microphonePermission?: string;
-    cameraPermission?: string;
+    microphonePermission?: string | false;
+    cameraPermission?: string | false;
   } | void
 > = (config, { microphonePermission, cameraPermission } = {}) => {
-  if (!config.ios) config.ios = {};
-  if (!config.ios.infoPlist) config.ios.infoPlist = {};
-  config.ios.infoPlist.NSCameraUsageDescription =
-    cameraPermission || config.ios.infoPlist.NSCameraUsageDescription || CAMERA_USAGE;
-  config.ios.infoPlist.NSMicrophoneUsageDescription =
-    microphonePermission || config.ios.infoPlist.NSMicrophoneUsageDescription || MICROPHONE_USAGE;
+  IOSConfig.Permissions.createPermissionsPlugin({
+    NSCameraUsageDescription: 'Allow $(PRODUCT_NAME) to access your camera',
+    NSMicrophoneUsageDescription: 'Allow $(PRODUCT_NAME) to access your microphone',
+  })(config, {
+    NSCameraUsageDescription: cameraPermission,
+    NSMicrophoneUsageDescription: microphonePermission,
+  });
 
   return AndroidConfig.Permissions.withPermissions(config, ['android.permission.CAMERA']);
 };