+++ packages/expo-image-picker/plugin/src/__tests__/withImagePicker-test.ts
@@ -1,7 +1,7 @@
 import { readAndroidManifestAsync } from '@expo/config-plugins/build/android/Manifest';
 import path from 'path';
 
-import { setImagePickerInfoPlist, withAndroidImagePickerPermissions } from '../withImagePicker';
+import { withAndroidImagePickerPermissions } from '../withImagePicker';
 
 function getMockAndroidManifest() {
   return readAndroidManifestAsync(
@@ -91,51 +91,5 @@ describe(withAndroidImagePickerPermissions, () => {
         application: expect.anything(),
       },
     });
-  });
-});
-
-describe(setImagePickerInfoPlist, () => {
-  it(`adds defaults to the plist`, () => {
-    expect(setImagePickerInfoPlist({}, {})).toStrictEqual({
-      NSCameraUsageDescription: 'Allow $(PRODUCT_NAME) to access your camera',
-      NSMicrophoneUsageDescription: 'Allow $(PRODUCT_NAME) to access your microphone',
-      NSPhotoLibraryUsageDescription: 'Allow $(PRODUCT_NAME) to access your photos',
-    });
-  });
-
-  it(`uses custom messages`, () => {
-    expect(
-      setImagePickerInfoPlist(
-        {
-          NSPhotoLibraryUsageDescription: 'foobar',
-          NSCameraUsageDescription: 'foobar',
-          NSMicrophoneUsageDescription: 'foobar',
-        },
-        {
-          cameraPermission: 'yolo',
-        }
-      )
-    ).toStrictEqual({
-      NSCameraUsageDescription: 'yolo',
-      NSMicrophoneUsageDescription: 'foobar',
-      NSPhotoLibraryUsageDescription: 'foobar',
-    });
-  });
-
-  it(`disables defaults explicitly`, () => {
-    expect(
-      setImagePickerInfoPlist(
-        {
-          NSPhotoLibraryUsageDescription: 'foobar',
-          NSCameraUsageDescription: 'foobar',
-          NSMicrophoneUsageDescription: 'foobar',
-        },
-        {
-          cameraPermission: false,
-          microphonePermission: false,
-          photosPermission: false,
-        }
-      )
-    ).toStrictEqual({});
   });
 });