+++ packages/expo-image-picker/CHANGELOG.md
@@ -6,6 +6,14 @@
 
 ### 🎉 New features
 
+### 🐛 Bug fixes
+
+### 💡 Others
+
+## 15.0.0 — 2024-04-18
+
+### 🎉 New features
+
 - Add ability to disable permissions in config plugin by passing `false` instead of permission messages. ([#28107](https://github.com/expo/expo/pull/28107) by [@EvanBacon](https://github.com/EvanBacon))
 
 ### 🐛 Bug fixes